Some changes to orbit
The moon orbits elipitcally because physics, it's not a major problem
and eliptical orbits are cool anyway. If you set the period of an orbit
too low then the speed will cause the transform to appear in the wrong
place in unity, I don't know why it happens but as far as I can tell
it's not actually affecting anything, it's just a visual bug.