Battle movement prototype (Gamepad)
On Dualshock the controls are as follows, X to accelerate, Square to
right your orientation, Circle to decelerate/reverse,  Triangle for
"hand brake", left analogue stick to change the pitch/yaw and L1, R1 for
rolling the ship.

For an Xbox controller just switch all the button names with the closest
available button. It's shouldn't be fundamentally different.

I took out the altitude changing method as I couldn't see much value in
it and there are only so many buttons on the pad without using
combinations (which I don't want to do).

I think it feels pretty good, so tell me what you think.

Also we're not using physics based rotation cause it feels too sluggish.

Stay tuned for the mouse update.