Made a bunch of prefabs
I also tweaked the movement (again), it's a bit slower now. Not sure if
we should keep it that way but if we're going to eventually implement
ship upgrades then it cant start out too strong.

I think the stats right now are a good balance but only when considering
upgrades. See what you guys think.