Simple AI for enemy, planet interface started
It will pursue the player, if it's on a collision course it will try and
stop. It's a bit stupid though, if it's too close already it will turn
toward the player and collide even though it has no hope of stopping in
time. Though this shouldn't be a big deal unless you're completely
stationary.