Fixes to dialogue system, base mission scene
Started working on demo mission too, though it's just dialogue 'cause I
was working on that system at the time.

ETA on demo mission build is one or two days if I get the time to do
lots of work and there are no tricky bugs. Even in that time frame it
will be really unpolished so realistically it'll probably end up taking
another week.